Move support bundle collection into a shared crate
Lifts `nexus/src/app/background/tasks/support_bundle/` (the mechanism
layer) into a new top-level crate `support-bundle-collection` so that
both Nexus and omdb can call it. No logic changes; pure relocation
plus import rewriting.
