+++ dev-tools/ls-apis/tests/api_dependencies.out
@@ -49,7 +49,7 @@ Management Gateway Service (client: gateway-client)
     consumed by: dpd (dendrite/dpd) via 1 path
     consumed by: lldpd (lldp/lldpd) via 1 path
     consumed by: mgd (maghemite/mgd) via 1 path
-    consumed by: omicron-nexus (omicron/nexus) via 5 paths
+    consumed by: omicron-nexus (omicron/nexus) via 6 paths
     consumed by: omicron-sled-agent (omicron/sled-agent) via 1 path
     consumed by: wicketd (omicron/wicketd) via 3 paths
 
@@ -96,7 +96,7 @@ Repo Depot API (client: repo-depot-client)
     consumed by: omicron-sled-agent (omicron/sled-agent) via 1 path
 
 Sled Agent (client: sled-agent-client)
-    consumed by: omicron-nexus (omicron/nexus) via 8 paths
+    consumed by: omicron-nexus (omicron/nexus) via 9 paths
 
 Wicketd (client: wicketd-client)
