+++ nexus/src/app/background/tasks/mod.rs
@@ -49,7 +49,6 @@ pub mod region_snapshot_replacement_step;
 pub mod saga_recovery;
 pub mod service_firewall_rules;
 pub mod session_cleanup;
-pub mod support_bundle;
 pub mod support_bundle_collector;
 pub mod sync_service_zone_nat;
 pub mod sync_switch_configuration;
