+++ support-bundle-collection/src/steps/sled_cubby.rs
@@ -4,9 +4,9 @@
 
 //! Collect sled cubby information for support bundles
 
-use crate::app::background::tasks::support_bundle::cache::Cache;
-use crate::app::background::tasks::support_bundle::collection::BundleCollection;
-use crate::app::background::tasks::support_bundle::step::CollectionStepOutput;
+use crate::cache::Cache;
+use crate::collection::BundleCollection;
+use crate::step::CollectionStepOutput;
 
 use anyhow::Context;
 use anyhow::bail;
