+++ support-bundle-collection/src/steps/sp_dumps.rs
@@ -4,11 +4,11 @@
 
 //! Collect SP task dumps for support bundles
 
-use crate::app::background::tasks::support_bundle::cache::Cache;
-use crate::app::background::tasks::support_bundle::collection::BundleCollection;
-use crate::app::background::tasks::support_bundle::step::CollectionStep;
-use crate::app::background::tasks::support_bundle::step::CollectionStepOutput;
-use crate::app::background::tasks::support_bundle::steps;
+use crate::cache::Cache;
+use crate::collection::BundleCollection;
+use crate::step::CollectionStep;
+use crate::step::CollectionStepOutput;
+use crate::steps;
 
 use anyhow::Context;
 use anyhow::bail;
